+++ tests/into_app.rs
